requirements.txt
allure-pytest==2.13.2
pytest==7.4.0
allure-python-commons==2.13.2
pytest-check== 2.5.3
